add swagger definitions to validationresponse
