--- apps/services/form-system/src/app/dataTypes/validationResponse.model.ts
@@ -1,22 +1,0 @@
-import { LanguageType } from './languageType.model'
-
-export class ScreenValidationResponse {
-  isValid!: boolean
-  screenId!: string
-  applicationId!: string
-  message!: LanguageType
-  fields!: FieldValidation[]
-}
-
-export class FieldValidation {
-  isValid!: boolean
-  fieldId!: string
-  message!: LanguageType
-  values!: ValueValidation[]
-}
-
-export class ValueValidation {
-  isValid!: boolean
-  valueId!: string
-  message!: LanguageType
-}
