+++ apps/services/form-system/src/app/modules/applications/applications.service.ts
@@ -24,7 +24,7 @@ import { Organization } from '../organizations/models/organization.model'
 import { ServiceManager } from '../services/service.manager'
 import { ApplicationEvent } from './models/applicationEvent.model'
 import { ApplicationListDto } from './models/dto/applicationList.dto'
-import { ScreenValidationResponse } from '../../dataTypes/validationResponse.model'
+import { ScreenValidationResponse } from '@island.is/form-system/shared'
 import { User } from '@island.is/auth-nest-tools'
 import { Applicant } from '../applicants/models/applicant.model'
 import { FormApplicantType } from '../formApplicantTypes/models/formApplicantType.model'
