+++ libs/form-system/src/dto/validationResponse.model.ts
@@ -1,0 +1,60 @@
+import { ApiProperty } from '@nestjs/swagger'
+import { LanguageType } from '../../../../apps/services/form-system/src/app/dataTypes/languageType.model'
+import { Field, ObjectType } from '@nestjs/graphql'
+
+@ObjectType('FormSystemValueValidation')
+export class ValueValidation {
+  @ApiProperty()
+  @Field()
+  isValid!: boolean
+
+  @ApiProperty()
+  @Field()
+  valueId!: string
+
+  @ApiProperty({ type: LanguageType })
+  @Field(() => LanguageType)
+  message!: LanguageType
+}
+
+@ObjectType('FormSystemFieldValidation')
+export class FieldValidation {
+  @ApiProperty()
+  @Field()
+  isValid!: boolean
+
+  @ApiProperty()
+  @Field()
+  fieldId!: string
+
+  @ApiProperty({ type: LanguageType })
+  @Field(() => LanguageType)
+  message!: LanguageType
+
+  @ApiProperty({ type: [ValueValidation] })
+  @Field(() => [ValueValidation])
+  values!: ValueValidation[]
+}
+
+@ObjectType('FormSystemScreenValidationResponse')
+export class ScreenValidationResponse {
+  @ApiProperty()
+  @Field()
+  isValid!: boolean
+
+  @ApiProperty()
+  @Field()
+  screenId!: string
+
+  @ApiProperty()
+  @Field()
+  applicationId!: string
+
+  @ApiProperty({ type: LanguageType })
+  @Field(() => LanguageType)
+  message!: LanguageType
+
+  @ApiProperty({ type: [FieldValidation] })
+  @Field(() => [FieldValidation])
+  fields!: FieldValidation[]
+}
