+++ libs/form-system/src/index.ts
@@ -10,3 +10,4 @@ export * from './enums/timeIntervals.enum'
 export * from './enums/urlMethods.enum'
 export * from './enums/urlTypes.enum'
 export * from './dto/updateForm.dto'
+export * from './dto/validationResponse.model'
